mil1553: New driver for the cbmia card to replace Yuris driver
This driver is a pure cbmia driver it knows nothing about RTIs power supplies or
the quick data protocol.

Three libray layers sit above this driver...

libmil1553 which is just a bunch of wrappers around the driver ioctl calls
librti which knows how to treat RTSs of various types and provides basic access to RTI registers
libquick which knows about the quick data protocol used by power supplies and other end user equipment

The driver provides one transaction queue for each bus controller that a client may place items on.
These items are executed under interrupt control and the results are placed back on the calling clients queue.
It is thus possible to perform simultaneous transactions on multiple bus controllers. However
prescise knowledge of the RTI cards is required if actions are carried out simultaneously on RTIs on the
same bus controller. This knowledge must be incorporated in library logic and in the client code logic,
in this case a simple command response logic is inadequate.

Signed-off-by: Julian Lewis <julian.lewis@cern.ch>