libquick: This library implements a simplified quick data protocol.
The quick data protocol is understood directly by the end user equipment (protocol #31)
In particular the G64 Power supplies. As all power-supply messages can be stored in a single
32 word cbmia buffer, the library does not implement the so called multiple telegram feature.
Thus although the QDP (Quick Data Protocol) permits packets up to 240 bytes long, in this
implementation no packet can be larger than 64 bytes. This restriction corresponds exactly
to the one imposed by the power supply itself. It would be a trivial task to implement multiple
telegrams, but it is not needed.

Although the library implements "raw" access in which the byte stream is sent directly to the hardware
it is a very very bad idea to use it on power supplies. This is because the way a raw byte stream apears in the
Power supply memory depends on the way the driver performs the buffer writes and on what seems to be
very inconsistent treatments in the power supply itself. If you insist on using raw access here are the rules.

Word order is little endian except within floating point numbers when writing.
The byte order is big endian except when the bytes are fields in a structure.
The byte order when reading back floats is little endian.
The alignement rules affect the sizeof structures.
To undersand the full horror refer to the serialization routines in the library.

Conclusion use pure native structures and transfer them via the send and recieve functions
that will serialize your data correctly when sending and receiving from the power supply.

Signed-off-by: Julian Lewis <julian.lewis@cern.ch>