mil1553: fix buffer overflow in client queue
The bug manifests with versions > 1.69 of the CBMIA firmware;
after some operations that require taking the bc_lock, the driver
attempts to take it twice, because the bc_lock field in the client
context is set to zero by the buffer overflow.

The overflow happens at lines 970ff of mil1553, that write to the
Rx buffer of the affected client. In particular, it occurs whenever
the received item fills the queue completely, and the received
packet uses up a full bunch of 32 words of the Rx buffer.

While this fixes the problem temporarily, the logic of the reception
should be corrected to avoid the overrun. In particular, it should be
taken into account that the word count can be 0 when there are 32
words to read OR when there is nothing to read (e.g. as a reply
to a SEND request).