mil1553: suppress kernel thread to keep rti mask up
The mask of up_rtis is used only in two places:

* the test program, to know what RTIs are alive
* at send_items, to avoid sending to a non-existent RTI

The mask of up RTIs is kept up to date by means of a
per-BC kernel thread that pings all the RTIs of a BC
and sets only the bits of the RTIs that reply. The thread
is woken up each time an interrupt comes and the thread
pings if it timed out waiting for an interrupt...

This change removes that logic from the driver: the up_rtis
mask is kept by updating to 1 the bit of each RTI that effectively
interrupts, and updating to 0 the same bit when an RTI reply times
out. SENDs to RTIs that are supposed not to exist are done
anyway: this keeps the mask updated.

To keep the test program happy, get_up_rtis forces a ping.