mil1553: remove offending spinlock grab
Grabbing mdev->lock to "protect" ping_rtis is incorrect.
To begin with, each individual RTI ping is a single transaction,
there is no need to lock anything during 30 of them.
Second, mutex grabs and wait_queues will happend with this
spinlock grabbed. The individual RTI ping itself (a do_start_tx)
is the real critical section, and it's already mutexed.