mil1553: define shorter ISR
The ISR should just note the fact that a reply has been correctly
received (or not) and leave porcess context the chore of copying
the data in exclusive mode (in which it will be anyway).

This minimizes the possibility of missing interrupts in high-contention
situations when there are shared lines. Moreover, it paves
the way for removal of all the queueing affairs, as seen in the
next commits

consequence: remove old ISR routine