mil1553: do not wait (that much) on the int_busy flag
An invariant of do_start_tx is that we enter with the bc non-busy
and without any pending int, and we exit the same way. These
invariants cannot be fulfilled if int handling takes forever,
which means more than a few milliseconds.

This changeset lightens the enforcement of these pre- and post-
conditions so that 1. the code is clea[rn]er, 2. more correct, and
3. does not impact performance that much.

At the outset, we just wait for the int_busy flag to clear. Four
msecs (INT_MISSING_TIMEOUT) is enough, a bit too much, and it is
interesting that the int_busy flag is still on while we never
get the dreadful HSTAT_BUSY_BIT error message. This points to
the host taking forever to handle some interrupts (!!!?????).

After that, the Tx write/interrupt cycle/wait for completion
should never take more than CBMIA_INT_TIMEOUT (2ms), but it
indeed does some times. So it seems that the isr never gets
called in that case, as the interrupt count of the bc differs from
the Tx count by exactly the same number of CBMIA_INT_TIMEOUT
events registered in the logs.

I do not fully understand what happens, but it is one of these:
a. either a missed interrupt (not by the CBMIA, who delivers
  flawlessly without contention), or
b. something causing the kernel to take more than 1sec to
  complete the ISR (unlikely).

Where the hell are those interrupts going to?