mil1553: correct bogus timeout interpretation
Checking for completion of the Tx - interrupt - Rx cycle
was done on the timeout return value of wait_queue*. Unfortunately,
as the granularity of the timer is a jiffie (1ms in our case) and
some times less than 1 jiffie is left, a 0 return value was
interpreted as a missing interrupt, which it is not.

This fix verifies instead that the operation has completed,
which is the Right Thing.